[core] Upgrade GetHitRate line to use int16 instead of int8 on param
May not currently be a bug, but >128 acc bonus in a calculation is easy

+++ src/map/entities/battleentity.h
@@ -558,7 +558,7 @@ public:
     uint16 CHR();
     uint16 DEF();
     uint16 ATT(SLOTTYPE slot);
-    uint16 ACC(uint8 attackNumber, uint8 offsetAccuracy);
+    uint16 ACC(uint8 attackNumber, uint16 offsetAccuracy);
     uint16 EVA();
     uint16 RATT(uint8 skill, uint16 bonusSkill = 0);
     uint16 RACC(uint8 skill, uint16 bonusSkill = 0);
